ESP32 WEMOS D1 R32, MPU9250, Wi-Fi, ESP- NOW, encrypted communication, graphical user interface, 3- axis G-force meter
A prototype gives the automobile driver option to read current and maximal G force in all three axes and to read the angle at which the vehicle is. The system consists of two parts. The first part is visible to the driver, and it consists of ESP32 and TFT display. The second part is in the trunk, and it consists of ESP32 and sensor MPU9250. 5 V power supply powers both units. Modules communicate over Wi-Fi, using ESP-NOW technology. While driving through corners, all passengers can get an interactive view of G force. The system uses encrypted communication. The price of hardware can go to about 20 $, which makes it suitable for educational purposes. Also, this fits for the training of engineers in developing countries.