IIoT, embedded programming, frontend development, ESP32, ESP-NOW, HTTPS, responsive  design, SPIFFS, C++, JavaScript, HTML5, CSS3, Highcharts
Prior to IIoT (The Industrial Internet of Things), embedded programming and frontend development 
didn't even found themselves in the same sentence. Hardware that enables seamless integration of these two vast 
areas is the Espressif ESP32 MCU (MicroController Unit). Once numerous devices were connected to the internet, 
the security of data became the focal point. ESP-NOW technology is Espressif’s proprietary solution for wireless 
data transfer. It can be encrypted and it enables secure communication between multiple ESP32's. HTTPS 
(Hypertext Transfer Protocol Secure) is a protocol that increases security on the internet. 
This article provides a guide for configuring a development environment for ESP32, ESP-NOW example, HTTPS 
server example, and programming practice intended for error handling. In addition, it offers a responsive example of a 
graphical user interface of the IIoT device. At the moment, there are numerous possible paths in embedded and 
frontend programming. This paper follows the path using ESP32 as the hardware, and C++ for the firmware.
JavaScript, HTML5, and CSS3 are unavoidable parts of modern industrial devices. Hence, the article provides an 
example of using JavaScript Highcharts library. This combination of hardware and software is the solution that costs 
less than 10$, which makes it acceptable in countries under development. Highcharts library is proprietary, and for 
educational purposes, it is used under Creative Commons (CC) Attribution-Non-Commercial license.